Campaigns: Move utils to separate package
Refs SHUUP-3008
