+++ shuup/campaigns/models/contact_group_sales_ranges.py
@@ -11,7 +11,7 @@ from django.db import models
 from django.db.models import Q
 from django.utils.translation import ugettext_lazy as _
 
-from shuup.campaigns.utils import get_contacts_in_sales_range
+from shuup.campaigns.utils.sales_range import get_contacts_in_sales_range
 from shuup.core.fields import MoneyValueField
 from shuup.core.models import ContactGroup, Shop
 from shuup.core.models._contacts import PROTECTED_CONTACT_GROUP_IDENTIFIERS
